\name{CreateRcode}
\alias{CreateRcode}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CreateRcode(codeFile, projName, mainDir, sections = NULL)}

\arguments{
	\item{codeFile }{FILL UP}

	\item{projName }{FILL UP}

	\item{mainDir }{FILL UP}

	\item{sections }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{AUTHOR(S) NAME(S) \email{authoremail}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

