\name{CreateRproj-package}
\alias{CreateRproj-package}
\alias{CreateRproj}

\docType{package}

\title{
 Create directory structures and ancillary files for R packages and R projects.
}

\description{
 This package provides functions to create the directory structure, description and namespace files for R packages, as well as directory structure for R projects. It also provides functions to create templates of R script files.
}

\details{
\tabular{ll}{
	Package: \tab CreateRproj \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2021-10-27\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}
