\name{CreatePkgDescrip}
\alias{CreatePkgDescrip}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CreatePkgDescrip(pkgName, mainDir)}

\arguments{
	\item{pkgName }{FILL UP}

	\item{mainDir }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{AUTHOR(S) NAME(S) \email{authoremail}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

