\name{CreateRdFiles}
\alias{CreateRdFiles}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CreateRdFiles(pkgName, mainDir, codeFile)}

\arguments{
	\item{pkgName }{FILL UP}

	\item{mainDir }{FILL UP}

	\item{codeFile }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{AUTHOR(S) NAME(S) \email{authoremail}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

