\name{CreateRproj-package}
\alias{CreateRproj-package}
\alias{CreateRproj}

\docType{Package}

\title{
 FILL UP
}

\description{
 FILL UP
}

\details{
\tabular{ll}{
	Package: \tab CreateRproj \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2021-03-24\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}

FILL UP DETAILS 
}

\author{AUTHOR(S) NAME(S) \email{authoremail}}

\references{
FILL UP
}