\name{CreateRproj}
\alias{CreateRproj}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CreateRproj(projName, mainDir, git = FALSE, RstProj = FALSE)}

\arguments{
	\item{projName }{FILL UP}

	\item{mainDir }{FILL UP}

	\item{git }{FILL UP}

	\item{RstProj }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{AUTHOR(S) NAME(S) \email{authoremail}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

